requirements.txt
streamlit
streamlit_ace
psycopg2-binary
cryptography
pycryptodome
matplotlib